Price a quantization recipe's call sites in its format in the floor model
A recipe declares, per call site, the tensor-core format its FLOPs run in
and the bytes per element of its quantized operands (Pricing). The graph
costs those call sites with them, Invocation carries the format, and
tools.profiling.floor divides each call site by its format's datasheet
peak and observed rate. RTX 5090 gains the block-scaled MXFP8 rows: 2048
FLOP/cycle/SM, 838 TFLOP/s at boost, 990 TFLOP/s observed (measured
2026-09-23). The BF16 floor is unchanged (25.143 / 25.202 ms); Pi0.5 under
mxfp8-llm-ffn prices at 13.29 / 15.45 ms. The report records the constants
file relative to the project.

+++ src/flash_vla/runtime/ops.py
@@ -99,7 +99,7 @@ class OpSpec:
         if not self.outputs:
             raise ValueError(f"{self.name}: an op must write at least one output")
 
-    def cost(self, shapes: Shapes, itemsizes: Mapping[str, int]) -> Cost:
+    def cost(self, shapes: Shapes, itemsizes: Mapping[str, float]) -> Cost:
         """Minimal traffic and math of one invocation with these argument shapes.
 
         Reads are every tensor argument that is not an output, plus the
